chore: Migrate reusabel R&M mocks to /shared/mocks
