+++ src/cypress/fixtures/get-rick-and-morty-charater.mock.ts
@@ -1,25 +1,10 @@
+import { mockCharacterData } from 'shared/mocks';
+
 const characterCountAlias = 'characterCountQuery';
 const characterCountOperationName = 'GetRickAndMortyCharacterCount';
 
 const charactersByIdAlias = 'charactersByIdQuery';
 const charactersByIdOperationName = 'GetRickAndMortyCharactersById';
-
-const mockCharacterCountResponse = {
-  data: { characters: { info: { count: 1 } } },
-};
-
-const mockCharacterData = {
-  id: '1',
-  name: 'Rick Sanchez',
-  species: 'Human',
-  image: 'https://rickandmortyapi.com/api/character/avatar/1.jpeg',
-};
-
-const mockCharactersByIdResponse = {
-  data: {
-    charactersByIds: [mockCharacterData],
-  },
-};
 
 const { id, name, species } = mockCharacterData;
 const mockCharacterDescription = `${id} | ${name} | ${species}`;
@@ -30,7 +15,5 @@ export {
   charactersByIdAlias,
   charactersByIdOperationName,
   mockCharacterData,
-  mockCharacterCountResponse,
-  mockCharactersByIdResponse,
   mockCharacterDescription,
 };
