+++ src/shared/mocks/index.ts
@@ -1,0 +1,11 @@
+import {
+  mockCharacterCountResponse,
+  mockCharacterData,
+  mockCharactersByIdResponse,
+} from './rickAndMortyCharacterMocks';
+
+export {
+  mockCharacterCountResponse,
+  mockCharacterData,
+  mockCharactersByIdResponse,
+};
